extract PlanBox component
