+++ lib/plausible_web/components/billing/plan_benefits.ex
@@ -24,12 +24,9 @@ defmodule PlausibleWeb.Components.Billing.PlanBenefits do
   """
   def render(assigns) do
     ~H"""
-    <ul
-      role="list"
-      class="mt-8 space-y-3 text-sm leading-6 xl:mt-10 text-gray-300 dark:text-gray-100"
-    >
+    <ul role="list" class="mt-8 space-y-3 text-sm leading-6 xl:mt-10 text-gray-300 dark:text-gray-100">
       <li :for={benefit <- @benefits} class="flex gap-x-3">
-        <Heroicons.check class={"h-6 w-5 text-indigo-600 dark:text-green-600"} />
+        <Heroicons.check class="h-6 w-5 text-indigo-600 dark:text-green-600" />
         <%= if is_binary(benefit), do: benefit, else: benefit.(assigns) %>
       </li>
     </ul>
